+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/BlockEntityFramingTable.java
@@ -164,22 +164,6 @@ public class BlockEntityFramingTable extends BaseBlockEntity implements Nameable
         if (ModCommonConfig.INSTANCE.DRAWERS.framed.enforceOpaqueMaterials.get()) {
             if (!state.canOcclude())
                 return false;
-
-            try {
-                // Will always throw unless overridden, which usually means it's a block that we don't
-                // want to be a valid material
-                if (state.getLightBlock(null, null) < 15)
-                    return false;
-            } catch (Exception e) { }
-
-            try {
-                if (!Block.isShapeFullBlock(state.getOcclusionShape(null, null)))
-                    return false;
-                if (state.propagatesSkylightDown(null, null))
-                    return false;
-            } catch (Exception e) {
-                return false;
-            }
         }
 
         if (MaterialBlacklist.INSTANCE.isBlacklisted(stack))