+++ forge/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/PlatformBlockEntityDrawersStandard.java
@@ -23,6 +23,12 @@ public abstract class PlatformBlockEntityDrawersStandard extends BlockEntityDraw
             super(pos, state);
         }
 
+        @Override
+        public void onLoad () {
+            super.onLoad();
+            onEntityLoad();
+        }
+
         @NotNull
         @Override
         public ModelData getModelData () {
@@ -46,6 +52,12 @@ public abstract class PlatformBlockEntityDrawersStandard extends BlockEntityDraw
             super(pos, state);
         }
 
+        @Override
+        public void onLoad () {
+            super.onLoad();
+            onEntityLoad();
+        }
+
         @NotNull
         @Override
         public ModelData getModelData () {
@@ -67,6 +79,12 @@ public abstract class PlatformBlockEntityDrawersStandard extends BlockEntityDraw
     public static class Slot4 extends BlockEntityDrawersStandard.Slot4 {
         public Slot4 (BlockPos pos, BlockState state) {
             super(pos, state);
+        }
+
+        @Override
+        public void onLoad () {
+            super.onLoad();
+            onEntityLoad();
         }
 
         @NotNull