+++ common/src/main/java/com/jaquadro/minecraft/storagedrawers/block/tile/tiledata/UpgradeData.java
@@ -87,8 +87,8 @@ public class UpgradeData extends BlockEntityDataShim
 
             ItemUpgradeStorage target = (ItemUpgradeStorage)prevUpgrade.getItem();
             ItemUpgradeStorage source = (ItemUpgradeStorage)upgrade.getItem();
-            if (source.level.getLevel() <= target.level.getLevel())
-                return false;
+            //if (source.level.getLevel() <= target.level.getLevel())
+            //    return false;
         }
 
         upgrades[slot] = ItemStack.EMPTY;